# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/health": {
            "get": {
                "description": "Returns service health and runtime information",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.HealthResponse"
                        }
                    }
                }
            }
        },
        "/session/create": {
            "post": {
                "description": "Creates a new session with the given TTL and session timeout",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "session"
                ],
                "summary": "Create a new session",
                "parameters": [
                    {
                        "description": "Session create payload",
                        "name": "session",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.SessionCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.SessionCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/session/terminate": {
            "post": {
                "description": "Terminates an active session by its token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "session"
                ],
                "summary": "Terminate a session",
                "parameters": [
                    {
                        "description": "Session terminate payload",
                        "name": "session",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.SessionTerminateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully terminated session.",
                        "schema": {
                            "$ref": "#/definitions/main.SessionTerminateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/session/update": {
            "post": {
                "description": "Updates the location information (longitude \u0026 latitude) of an active session identified by its token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "session"
                ],
                "summary": "Update a session",
                "parameters": [
                    {
                        "description": "Session update payload",
                        "name": "session",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.SessionUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Session updated.",
                        "schema": {
                            "$ref": "#/definitions/main.SessionUpdateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid json"
                }
            }
        },
        "main.HealthResponse": {
            "type": "object",
            "properties": {
                "runtime": {
                    "type": "string",
                    "example": "38271.133967079s"
                },
                "status": {
                    "type": "string",
                    "example": "OK"
                }
            }
        },
        "main.SessionCreateRequest": {
            "type": "object",
            "properties": {
                "session_timeout": {
                    "type": "integer",
                    "example": 7200
                },
                "ttl": {
                    "type": "integer",
                    "example": 3600
                }
            }
        },
        "main.SessionCreateResponse": {
            "type": "object",
            "properties": {
                "parameters": {
                    "$ref": "#/definitions/main.SessionCreateRequest"
                },
                "token": {
                    "type": "string",
                    "example": "3A9N2O"
                }
            }
        },
        "main.SessionTerminateRequest": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "3A9N2O"
                }
            }
        },
        "main.SessionTerminateResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "successfully terminated session."
                },
                "token": {
                    "type": "string",
                    "example": "3A9N2O"
                }
            }
        },
        "main.SessionUpdateRequest": {
            "type": "object",
            "properties": {
                "latitude": {
                    "type": "number",
                    "example": 8.385259
                },
                "longitude": {
                    "type": "number",
                    "example": 49.026598
                },
                "token": {
                    "type": "string",
                    "example": "3A9N2O"
                }
            }
        },
        "main.SessionUpdateResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "successfully updated session."
                },
                "token": {
                    "type": "string",
                    "example": "3A9N2O"
                }
            }
        }
    }
}